--- a/perfmetrics/scripts/presubmit_test/pr_perf_test/build.sh
+++ b/perfmetrics/scripts/presubmit_test/pr_perf_test/build.sh
@@ -69,18 +69,7 @@ function install_requirements() {
   pip install google-cloud-vision
   pip install google-api-python-client
   pip install prettytable
-  # install libaio as fio has a dependency on libaio
-  sudo apt-get install libaio-dev
-  # We are building fio from source because of issue: https://github.com/axboe/fio/issues/1640.
-  # The fix is not currently released in a package as of 20th Oct, 2023.
-  # TODO: install fio via package when release > 3.35 is available.
-  echo Installing fio
-  sudo rm -rf "${KOKORO_ARTIFACTS_DIR}/github/fio"
-  git clone https://github.com/axboe/fio.git "${KOKORO_ARTIFACTS_DIR}/github/fio"
-  cd  "${KOKORO_ARTIFACTS_DIR}/github/fio" && \
-  git checkout c5d8ce3fc736210ded83b126c71e3225c7ffd7c9 && \
-  ./configure && make && sudo make install
-  fio --version
+  "${KOKORO_ARTIFACTS_DIR}"/github/gcsfuse/perfmetrics/scripts/fio/install_fio.sh
   cd "${KOKORO_ARTIFACTS_DIR}/github/gcsfuse"
 }