--- a/perfmetrics/scripts/run_load_test_and_fetch_metrics.sh
+++ b/perfmetrics/scripts/run_load_test_and_fetch_metrics.sh
@@ -17,17 +17,7 @@ set -e
 
 echo "Installing pip"
 sudo apt-get install pip -y
-echo "Installing fio"
-# install libaio as fio has a dependency on libaio
-sudo apt-get install libaio-dev
-# We are building fio from source because of issue: https://github.com/axboe/fio/issues/1640.
-# The fix is not currently released in a package as of 20th Oct, 2023.
-# TODO: install fio via package when release > 3.35 is available.
-sudo rm -rf "${KOKORO_ARTIFACTS_DIR}/github/fio"
-git clone https://github.com/axboe/fio.git "${KOKORO_ARTIFACTS_DIR}/github/fio"
-cd  "${KOKORO_ARTIFACTS_DIR}/github/fio" && \
-git checkout c5d8ce3fc736210ded83b126c71e3225c7ffd7c9 && \
-./configure && make && sudo make install
+"${KOKORO_ARTIFACTS_DIR}"/github/gcsfuse/perfmetrics/scripts/fio/install_fio.sh
 
 cd "${KOKORO_ARTIFACTS_DIR}/github/gcsfuse/perfmetrics/scripts"