### Description
Currently the commands for installing fio
is duplicated at two different places. This change merges them and puts it into a common script at
perfmetrics/scripts/fio/install_fio.sh and invokes it from both the places.

fio version 3.36 has been released, so we could directly installed that from the fio-3.36 package, but we need to build fio from source for the [fix in clat latency calculations](https://github.com/GoogleCloudPlatform/gcsfuse/pull/1503) .

This is precursor to https://github.com/GoogleCloudPlatform/gcsfuse/pull/1503.

### Link to the issue in case of a bug fix.
NA

### Testing details
1. Manual - NA
2. Unit tests - NA
3. Integration tests - NA