@gargnitingoogle 2023-11-19
the perf presubmit test has passed: [log](https://fusion2.corp.google.com/ci/kokoro/prod:gcsfuse%2Fgcp_ubuntu%2Fpresubmits%2Fperf_tests%2Fpresubmit/activity/840f67db-f420-47ef-96a0-fe8c2c9b0111/log).

showing results...
+--------+------------+--------------+------------+--------------+--------------+
| Branch | File Size  |   Read BW    |  Write BW  | RandRead BW  | RandWrite BW |
+--------+------------+--------------+------------+--------------+--------------+
| Master |  0.25MiB   | 493.99MiB/s  | 0.94MiB/s  |  34.99MiB/s  |   1.1MiB/s   |
|   PR   |  0.25MiB   | 445.37MiB/s  | 1.09MiB/s  |  33.46MiB/s  |   1.1MiB/s   |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 48.828MiB  | 4732.38MiB/s | 76.5MiB/s  | 1111.49MiB/s |  76.63MiB/s  |
|   PR   | 48.828MiB  | 4743.12MiB/s | 56.14MiB/s | 1171.55MiB/s |  79.56MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
| Master | 976.562MiB | 4661.63MiB/s | 29.03MiB/s |  663.2MiB/s  |  25.39MiB/s  |
|   PR   | 976.562MiB | 4556.66MiB/s | 31.49MiB/s | 753.69MiB/s  |  25.16MiB/s  |
|        |            |              |            |              |              |
|        |            |              |            |              |              |
+--------+------------+--------------+------------+--------------+--------------+